requirements.txt
onnxruntime==1.11.1

click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==2.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
onnxruntime==1.11.1
Pillow==9.1.1
protobuf==4.21.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
